removed join function
